This File Contains the Method to Parse XML File.
Parsing an xml file can be headache for some.
This class contains the method that can easily parse the xml file and get us the desired result.
Think of it during webservices.
When we Request webservice with some data it respond back generally in the form of the xml file. so when we receive the file we can take the xml response from the webservice and save it into and .xml file. After that we can parse the file to get the desired values from the xml File element.
For reference i have added one xml file.